# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PlayMoney API",
    "version": "1.0"
  },
  "components": {
    "schemas": {
...
      "id": {
        "type": "string"
      },
...
    },
    "parameters": {
...
      "id": {
        "schema": {
          "$ref": "#/components/schemas/id"
        },
        "required": true,
        "name": "id",
        "in": "path"
      },
...
    }
  },
  "paths": {
...
    "/v1/comments": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 5000
                  },
                  "parentId": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "entityType": {
                    "type": "string",
                    "enum": [
                      "MARKET",
                      "LIST"
                    ]
                  },
                  "entityId": {
                    "type": "string"
                  }
                },
                "required": [
                  "content",
                  "parentId",
                  "entityType",
                  "entityId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "entityType": {
                      "type": "string",
                      "enum": [
                        "MARKET",
                        "LIST"
                      ]
                    },
                    "id": {
                      "type": "string",
                      "format": "cuid"
                    },
                    "content": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 5000
                    },
                    "createdAt": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "updatedAt": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "edited": {
                      "type": "boolean"
                    },
                    "authorId": {
                      "type": "string"
                    },
                    "parentId": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "hidden": {
                      "type": "boolean"
                    },
                    "entityId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "entityType",
                    "id",
                    "content",
                    "createdAt",
                    "updatedAt",
                    "edited",
                    "authorId",
                    "parentId",
                    "hidden",
                    "entityId"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/comments/[id]/reaction": {
      "post": {
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "emoji": {
                    "type": "string"
                  }
                },
                "required": [
                  "emoji"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": [
                  {
                    "_def": {
                      "unknownKeys": "strip",
                      "catchall": {
                        "_def": {
                          "typeName": "ZodNever"
                        }
                      },
                      "typeName": "ZodObject"
                    },
                    "_cached": null
                  },
                  {
                    "_def": {
                      "unknownKeys": "strip",
                      "catchall": {
                        "_def": {
                          "typeName": "ZodNever"
                        }
                      },
                      "typeName": "ZodObject"
                    },
                    "_cached": null
                  }
                ]
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}